[feat] board, user, comment 사용자 정의 예외처리 세분화
- Service Interface 추가
